Hu - Put the content into <div> 'top' </div>
Add three links and icons to the social media of  Edible garden on the left of the top area.
Icons are using the free Font Awesome library.
Add the search bar to the right area.
Match the same style on the official website of botanical garden.
The search bar is just a place hold.